cmd/frontman: refactor more code for testing
